refactor: VendingMachine.ts 경로 이동
