+++ src/__test__/itemInputValidation.test.ts
@@ -1,4 +1,4 @@
-import VendingMachine from '../ts/VendingMachine';
+import VendingMachine from '../ts/domain/VendingMachine';
 import { ITEM_ERROR_MESSAGE } from '../ts/constant/errorMessage';
 
 describe('상품 추가할 때, 입력값 유효성 확인', () => {
